직업 선택 훅 분리
